requirements-player-stats.txt
python-dotenv>=0.19.0
supabase>=1.0.3
ohmysportsfeedspy>=1.0.0
python-dateutil>=2.8.2
